The pol345.student Package

This package contains handouts for Marc Ratkovic and Will Lowe's version of POL 345.

Unpacking Handout Materials

Use get_handout, with a number to unpack handout materials in a local directory. For example:

get_handout(2)

This will unpack the second handout materials into a folder called 'handout2` in your current working directory. (Type getwd if you're not sure where that is. You can change it using setwd or through the graphical user interface).

Starting Again

If you want to start again on the same handout, you'll can unpack the handout materials again under a different name (get_handout won't overwrite an existing folder.) To do this use the newname argument. If you want your new copy to be called 'handout2-forreal', then use

get_handout(2, newname = "handout2-forreal")

Provided there's not already a folder of that name in your current working directory, you'll get a fresh set of handout materials unpacked there.

If you want to preview the questions in a handout materials without unpacking it into your local file system, use preview_handout.